BottleNumber -> Switch-Statement-code-Smell
To remove BottleNumber conditionals we can do two things:
Replace conditionals with Polymorphism -> Inheritance based
Replace conditionals with Strategy -> Composition based

For this case we will choose Polymophism.

To identify what is the concept which varies we take a look to boolean
conditionals.
We can see that we compare number == 1 and number == 0
We have three execution paths: 0, 1 and default.
So we will use three objects BottleNumber0, BottleNumber1 and BottleNumber
They all will share same interface (We ensure that with inheritance).